requirements.txt
numpy
pandas
Pillow
scikit-learn
torch
torchvision
matplotlib
seaborn
jupyter
